refactor(server): peel batch-retranscribe routes into routers/retranscribe.py (router-split, round-5 #51 / R5-25 PR20)
Ninth router peel. Moves the project-wide batch retranscribe group (Phase 9.6d)
into a leaf routers/retranscribe.py: POST /api/retranscribe-all, its /status
poller, the _run_retranscribe_all background worker, and RetranscribeAllRequest
(with its M23 ISO-639 language validator). The two-lock ordering is preserved
verbatim — acquire the retranscribe single-flight guard, then the shared H3
ingest slot (the batch runs whisper in-process, so it must hold the ingest slot
too or a concurrent /api/ingest double-loads whisper → OOM); the worker releases
both in its finally. Guard + slot are imported from state.py (the ONE shared
instance), never recreated.

Behaviour-preserving: paths/methods/scopes unchanged (SPA + CLI unaffected).
test_retranscribe_all.py's srv fixture repointed to routers.retranscribe (worker
+ same state.retranscribe guard). server.py 2967 → 2825. Boundary test pins
leaf-ness, route ownership, shared-guard identity, the moved validator, and
mounted+auth-guarded (401-not-404).
